backend: split generateMovementEdges
